% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_objects.R
\name{Task}
\alias{Task}
\title{TaskQueue API Objects 
Accesses a Google App Engine Pull Task Queue over REST.}
\usage{
Task(enqueueTimestamp = NULL, id = NULL, leaseTimestamp = NULL,
  payloadBase64 = NULL, queueName = NULL, retry_count = NULL,
  tag = NULL)
}
\arguments{
\item{enqueueTimestamp}{Time (in seconds since the epoch) at which the task was enqueued}

\item{id}{Name of the task}

\item{leaseTimestamp}{Time (in seconds since the epoch) at which the task lease will expire}

\item{payloadBase64}{A bag of bytes which is the task payload}

\item{queueName}{Name of the queue that the task is in}

\item{retry_count}{The number of leases applied to this task}

\item{tag}{Tag for the task, could be used later to lease tasks grouped by a specific tag}
}
\value{
Task object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:49:25
filename: /Users/mark/dev/R/autoGoogleAPI/googletaskqueuev1beta2.auto/R/taskqueue_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Task Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Task functions: \code{\link{tasks.insert}},
  \code{\link{tasks.patch}}, \code{\link{tasks.update}}
}

