% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_objects.R
\name{TaskQueue}
\alias{TaskQueue}
\title{TaskQueue Object}
\usage{
TaskQueue(TaskQueue.acl = NULL, TaskQueue.stats = NULL, acl = NULL,
  id = NULL, maxLeases = NULL, stats = NULL)
}
\arguments{
\item{TaskQueue.acl}{The \link{TaskQueue.acl} object or list of objects}

\item{TaskQueue.stats}{The \link{TaskQueue.stats} object or list of objects}

\item{acl}{ACLs that are applicable to this TaskQueue object}

\item{id}{Name of the taskqueue}

\item{maxLeases}{The number of times we should lease out tasks before giving up on them}

\item{stats}{Statistics for the TaskQueue object in question}
}
\value{
TaskQueue object
}
\description{
TaskQueue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TaskQueue functions: \code{\link{TaskQueue.acl}},
  \code{\link{TaskQueue.stats}}
}
