% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_objects.R
\name{TaskQueue.stats}
\alias{TaskQueue.stats}
\title{TaskQueue.stats Object}
\usage{
TaskQueue.stats(leasedLastHour = NULL, leasedLastMinute = NULL,
  oldestTask = NULL, totalTasks = NULL)
}
\arguments{
\item{leasedLastHour}{Number of tasks leased in the last hour}

\item{leasedLastMinute}{Number of tasks leased in the last minute}

\item{oldestTask}{The timestamp (in seconds since the epoch) of the oldest unfinished task}

\item{totalTasks}{Number of tasks in the queue}
}
\value{
TaskQueue.stats object
}
\description{
TaskQueue.stats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Statistics for the TaskQueue object in question.
}
\seealso{
Other TaskQueue functions: \code{\link{TaskQueue.acl}},
  \code{\link{TaskQueue}}
}
