% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_functions.R
\name{tasks.insert}
\alias{tasks.insert}
\title{Insert a new task in a TaskQueue}
\usage{
tasks.insert(Task, project, taskqueue)
}
\arguments{
\item{Task}{The \link{Task} object to pass to this method}

\item{project}{The project under which the queue lies}

\item{taskqueue}{The taskqueue to insert the task into}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/taskqueue
\item https://www.googleapis.com/auth/taskqueue.consumer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/taskqueue, https://www.googleapis.com/auth/taskqueue.consumer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/appengine/docs/python/taskqueue/rest}{Google Documentation}

Other Task functions: \code{\link{Task}},
  \code{\link{tasks.patch}}, \code{\link{tasks.update}}
}
