% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_functions.R
\name{tasks.patch}
\alias{tasks.patch}
\title{Update tasks that are leased out of a TaskQueue. This method supports patch semantics.}
\usage{
tasks.patch(Task, project, taskqueue, task, newLeaseSeconds)
}
\arguments{
\item{Task}{The \link{Task} object to pass to this method}

\item{project}{The project under which the queue lies}

\item{taskqueue}{}

\item{task}{}

\item{newLeaseSeconds}{The new lease in seconds}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/taskqueue
\item https://www.googleapis.com/auth/taskqueue.consumer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/taskqueue, https://www.googleapis.com/auth/taskqueue.consumer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/appengine/docs/python/taskqueue/rest}{Google Documentation}

Other Task functions: \code{\link{Task}},
  \code{\link{tasks.insert}}, \code{\link{tasks.update}}
}
