% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_objects.R
\name{Tasks}
\alias{Tasks}
\title{Tasks Object}
\usage{
Tasks(items = NULL)
}
\arguments{
\item{items}{The actual list of tasks returned as a result of the lease operation}
}
\value{
Tasks object
}
\description{
Tasks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

