% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_objects.R
\name{Tasks2}
\alias{Tasks2}
\title{Tasks2 Object}
\usage{
Tasks2(items = NULL)
}
\arguments{
\item{items}{The actual list of tasks currently active in the TaskQueue}
}
\value{
Tasks2 object
}
\description{
Tasks2 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

